soc: intel_adsp: Move adsp.c to common SOC area
Move adsp.c to common and clean makefiles.
