+++ soc/xtensa/intel_adsp/CMakeLists.txt
@@ -4,4 +4,3 @@
 # SPDX-License-Identifier: Apache-2.0
 
 add_subdirectory(common)
-add_subdirectory(${SOC_SERIES})
