--- soc/xtensa/intel_adsp/broadwell_adsp/CMakeLists.txt
@@ -1,5 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-
-zephyr_library()
-zephyr_library_link_libraries(intel_adsp_common)
-zephyr_library_include_directories(${ZEPHYR_BASE}/drivers)
